refactor(IT-Cotato#252): BoardFavoriteAppender, Deleter 병합
- 클래스명 쓰기작업을 하는 xxxManager로 변경
